requirements.txt
numpy
opencv-python
tensorflow
matplotlib
scikit-learn
pillow